requirements.txt
aniso8601==1.2.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12
Flask-Cors==3.0.2
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
gunicorn==19.6.0
idna==2.5
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
MySQL-python==1.2.5
pyotp==2.2.4
python-dateutil==2.6.0
python-ldap==2.4.28
pytz==2016.10
pyzmail==1.0.3
requests==2.18.1
six==1.10.0
SQLAlchemy==1.1.5
urllib3==1.21.1
Werkzeug==0.11.15
